# docs/swagger.yaml
definitions:
  dto.CredentialSignInDto:
    properties:
      deviceToken:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.CustomerSignup:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  dto.DriverSignup:
    type: object
...
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      kind:
        type: string
      message:
        type: string
    type: object
  dto.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.UserSignup:
    properties:
      address:
        type: string
      createdBy:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
...
info:
  contact: {}
paths:
  /auth/c/signin:
    post:
      consumes:
      - application/json
      description: A Customer authentication sign-in method
      parameters:
      - description: customer login
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Customer Sign-In
      tags:
      - Authentication - Customer
  /auth/c/signup:
    post:
      consumes:
      - application/json
      description: Customer Sign-Up
      parameters:
      - description: customer registration
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CustomerSignup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Customer Sign-Up
      tags:
      - Authentication - Customer
  /auth/d/signin:
    post:
      consumes:
      - application/json
      description: A Driver authentication sign-in method for mobile apps
      parameters:
      - description: driver login
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Driver Sign-In
      tags:
      - Authentication - Driver
  /auth/d/signup:
    post:
      consumes:
      - application/json
      description: Driver Sign-Up
      parameters:
      - description: driver registration
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DriverSignup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Driver Sign-Up
      tags:
      - Authentication - Driver
  /auth/i/signin:
    post:
      consumes:
      - application/json
      description: A Internal authentication sign-in method
      parameters:
      - description: internal login
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CredentialSignInDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Internal Sign-In
      tags:
      - Authentication - Internal
  /auth/i/signup:
    post:
      consumes:
      - application/json
      description: Internal Sign-Up
      parameters:
      - description: internal registration
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserSignup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Internal Sign-Up
      tags:
      - Authentication - Internal
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Universal Reset Password
      parameters:
      - description: universal reset password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DriverSignup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Universal Reset Password
      tags:
      - Authentication - Reset Password
...
swagger: "2.0"